CompletelyFreeAgent/requirements.txt
numpy==1.26.4
livekit-agents>=0.12.3
livekit-plugins-deepgram>=0.6.15
livekit-plugins-google>=0.8.1
livekit-plugins-openai[vertex]>=0.10.10
livekit-plugins-silero
livekit-plugins-rag>=0.2.3
livekit-plugins-elevenlabs
livekit-plugins-turn-detector
python-dotenv~=1.0
livekit-api
requests
livekit-plugins-rag
aiofile
llama-index-readers-file
llama-index-core
llama-index-embeddings-openai
torch
transformers
llama-index-llms-openai